+++ spec/rails_helper.rb
@@ -52,7 +52,7 @@ RSpec.configure do |config|
   #
   # The different available types are documented in the features, such as in
   # https://relishapp.com/rspec/rspec-rails/docs
-  config.infer_spec_type_from_file_location!
+  # config.infer_spec_type_from_file_location!
 
   # to be able to use route helpers in specs
   config.include Rails.application.routes.url_helpers