+++ spec/spec_helper.rb
@@ -74,6 +74,16 @@ RSpec.configure do |config|
     mocks.verify_partial_doubles = true
   end
 
+  config.include(FactoryGirl::Syntax::Methods,
+                 type: :services,
+                 file_path: %r{/spec\/services/}
+                )
+
+  config.include(FactoryGirl::Syntax::Methods,
+                 type: :query,
+                 file_path: %r{/spec\/queries/}
+                )
+
   if defined?(Bullet) && Bullet.enable?
     config.before(:each) do
       Bullet.start_request